added howto and report; changed creator.sh and destroyer.sh: now credentials can be got from shared file or from command line; same for client: credentials from CLI only if configure.sh is lauched

+++ src/consumer/init_infrastructure/python_source/lambda_delete_and_mark.py
@@ -14,5 +14,5 @@ def delete_and_mark(event, context):
                             'Value': 'False'
                         }]
                 })
-        else:
+        elif event[key] == 'del':
             client.delete_object(Bucket=MESSAGE_BUCKET_NAME, Key=key)
